rename editor.py.py
